Added skip functionality to unittest runner, made NCBI tests skip on macro test runner

+++ ptmscout_web/unittest.sh
@@ -2,17 +2,21 @@ if [ $# -eq 0 ]
 then
     for f in `find tests | grep "test_.*\.py$"`
     do
-        package=`echo $f | sed "s/\//./g" | sed "s/\(.*\)\.py/\1/"`
-        echo "Running tests: $package"
-        PYTHONPATH=/data/ptmscout/ptmscout_web /data/pyramid/bin/python -m unittest $package 2> tmp_test_output.txt
-        rc=$?
-        output=`cat tmp_test_output.txt`
-        rm tmp_test_output.txt
-        if [[ $rc != 0 ]]
+        proceed=`head -n1 $f | grep '# Skip' | wc -l`
+        if [ "$proceed" -eq "0" ]
         then
-            echo "Failed test..."
-            echo "$output"
-            exit $rc
+            package=`echo $f | sed "s/\//./g" | sed "s/\(.*\)\.py/\1/"`
+            echo "Running tests: $package"
+            PYTHONPATH=/data/ptmscout/ptmscout_web /data/pyramid/bin/python -m unittest $package 2> tmp_test_output.txt
+            rc=$?
+            output=`cat tmp_test_output.txt`
+            rm tmp_test_output.txt
+            if [[ $rc != 0 ]]
+            then
+                echo "Failed test..."
+                echo "$output"
+                exit $rc
+            fi
         fi
     done
